+++ lib/src/core/error/exceptions.dart
@@ -1,7 +1,13 @@
 class ServerException implements Exception {
+  final String url;
   final String? error;
+  final int? statusCode;
+  ServerException({required this.url, this.error, this.statusCode});
 
-  ServerException({this.error});
+  @override
+  String toString() {
+    return '$url => $error with: $statusCode';
+  }
 }
 
 class CacheException implements Exception {}