+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.2
+
+- Exported CreationMode enum
+
+## 0.1.1
+
+- Fixed aim children property, it wasn't nullable
+
+## 0.1.0
+
+- Implemented creation mode (standard, setLocationOnRedeem) 
+
 ## 0.0.2-nullsafety.3
 
 - Exposed Instrument class